feat: Move form to own page
