+++ src/components/LoginForm/index.tsx
@@ -51,7 +51,7 @@ function LoginFormFields() {
     <form onSubmit={handleSubmit(onSubmit)}>
       <Box
         sx={{
-          width: "min(90vw, 500px)",
+          width: "min(90vw, 700px)",
           display: "grid",
           gap: 1.5,
           gridTemplateColumns: "1fr",
